feat 36594086: added better structure of the bot
